- feat: Modifying services package, adding impl dir and moving implementation classes there.
- Order_Item entity created.

- Shipment_id modified.
