+++ src/main/java/com/alumnione/ecommerce/entity/OrderItem.java
@@ -1,0 +1,23 @@
+package com.alumnione.ecommerce.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.Date;
+
+@Entity
+@AllArgsConstructor @NoArgsConstructor
+@Getter @Setter
+@Table(name = "order_items")
+public class OrderItem {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long order_item_id;
+    private Date order_date;
+    private Float total_price;
+
+}
